// src/dto/user.dto.ts
export interface IUserInfo {
    userId?: string;
    firstName?: string;
    lastName?: string;
    city?: string;
    state?: string;
    country?: string;
    zipCode?: string;
    address?: string;
    phone?: string;
    otherEmail?: string;
    website?: string;
    gender?: string;
    description?: string;
    title?: string;
  };